requirements.txt
PyVisa
PyQt5
PyMySQL
numpy
pyserial
PyDAQmx
nidaqmx
loguru
